scripts/python/app/locale/{fr,it,ru}/LC_MESSAGES/NUT-Monitor.mo: regenerate due to changes in translation sources [#722]
Note that we keep verbatim @NUT PACKAGE_VERSION@ and @NUT_WEBSITE_BASE@
in text (About dialog) and substitute them in real-time. This allows us
to keep the document wording static, while certain strings depend on
build circumstances. Unlike other resources, this dialog is shared with
py2 code which is not localized in same manner as py3, and does not seem
to use formatting strings ("{0}" or "{}" sequences).

Signed-off-by: Jim Klimov <jimklimov+nut@gmail.com>