Makefile.am, docs/Makefile.am, NEWS.adoc: introduce a SPELLCHECK_INTERACTIVE_QUICK hack [#2871]
When we run `make spellcheck-interactive-quick` (from top dir) and one
of the files in a sub-directory with documents causes a change of the
dictionary, re-enter to run the probably-parallel `make spellcheck-quick`
(non-interactive) from the top directory again and update timestamps of
all text files that agree with the updated dictionary quickly. This may
save time on multi-CPU systems compared to a sequential walk of all texts.

Signed-off-by: Jim Klimov <jimklimov+nut@gmail.com>